# openapi.yaml
openapi: 3.0.0
info:
  title: Points API
  description: An API for modifying points as part of a campaign.
  version: 1.0.0
servers:
  - url: http://localhost:8080

paths:
  /campaigns/create:
    post:
      summary: Creates a new campaign.
      description: |
        Creates a new campaign, returning an id for the campaign and an API key.
        This API key provides authorization for endpoints related to that campaign.
      parameters:
        - in: header
          name: accept-version
          example: 1.0.0
          schema:
            type: string
          required: false
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  apiKey:
                    type: string
                    example: 20a067cd-6864-4293-95e4-b1350cd0c1fa
                  campaignId:
                    type: integer
                    example: 1234
                required:
                  - apiKey
                  - campaignId

  /campaigns/{campaignId}/points:
    get:
      summary: Get the number of points for a specific address.
      description: Optionally return only the points under a specified event name.
      security:
        - ApiKeyAuth: []
      parameters:
        - in: path
          name: campaignId
          schema:
            type: integer
          required: true
        - in: query
          name: address
          description: Ethereum address
          schema:
            type: string
          required: true
        - in: query
          name: event
          description: Name of the event to search for
          schema:
            type: string
          required: false
        - in: header
          name: accept-version
          example: 1.0.0
          schema:
            type: string
          required: false
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                type: number
              examples:
                example:
                  value: 1000
        400:
          $ref: "#/components/responses/ValidationError"
        401:
          $ref: "#/components/responses/UnauthorizedError"
        403:
          $ref: "#/components/responses/ForbiddenError"
        404:
          $ref: "#/components/responses/NotFoundError"
    put:
      summary: Add points to a specific address.
      description: |
        Returns the total number of points for the specified address.
        If an event name is specified, only count the points under that event.
      security:
        - ApiKeyAuth: []
      parameters:
        - in: path
          name: campaignId
          schema:
            type: integer
          required: true
        - in: header
          name: accept-version
          example: 1.0.0
          schema:
            type: string
          required: false
      requestBody:
        description: a
        required: true
        content:
          application/json:
            schema:
                type: object
                properties:
                  address:
                    type: string
                    example: 20a067cd-6864-4293-95e4-b1350cd0c1fa
                  event:
                    type: string
                    example: foo
                  points:
                    type: integer
                    example: 1000
                required:
                  - address
                  - event
                  - points
            examples:
              example:
                value:
                  address: 20a067cd-6864-4293-95e4-b1350cd0c1fa
                  event: foo
                  points: 1000

      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                type: number
        400:
          $ref: "#/components/responses/ValidationError"
        401:
          $ref: "#/components/responses/UnauthorizedError"
        403:
          $ref: "#/components/responses/ForbiddenError"
        404:
          $ref: "#/components/responses/NotFoundError"

components:
  responses:
    ForbiddenError:
      description: Resource is forbidden.
      content:
        application/json:
          schema:
            type: object
            properties:
              reason:
                type: string
                example: forbidden
              message:
                type: string
                example: API Key does not have access to this resource.
    NotFoundError:
      description: Resource not found.
      content:
        application/json:
          schema:
            type: object
            properties:
              reason:
                type: string
                example: not-found
              message:
                type: string
                example: No campaign found with id "1234".
    UnauthorizedError:
      description: Authentication is missing or invalid.
      content:
        application/json:
          schema:
            type: object
            properties:
              reason:
                type: string
                example: unauthorized
              message:
                type: string
                example: API Key is missing or invalid.
    ValidationError:
      description: Request validation failed.
      content:
        application/json:
          schema:
            type: object
            properties:
              reason:
                type: string
                example: validation
              message:
                type: string
                example: Request validation failed.
              errors:
                type: array
                example: ['Address must start with "0x".']
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-KEY